Refactor RenderingHandler into view_base.
