+++ gubernator/view_base.py
@@ -1,0 +1,56 @@
+# Copyright 2016 The Kubernetes Authors All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import os
+import re
+
+import jinja2
+import webapp2
+import yaml
+
+import filters as jinja_filters
+
+from google.appengine.api import urlfetch
+
+BUCKET_WHITELIST = {
+    re.match(r'gs://([^/]+)', path).group(1)
+    for path in yaml.load(open("buckets.yaml"))
+}
+
+JINJA_ENVIRONMENT = jinja2.Environment(
+    loader=jinja2.FileSystemLoader(os.path.dirname(__file__) + '/templates'),
+    extensions=['jinja2.ext.autoescape'],
+    trim_blocks=True,
+    autoescape=True)
+JINJA_ENVIRONMENT.line_statement_prefix = '%'
+jinja_filters.register(JINJA_ENVIRONMENT.filters)
+
+class RenderingHandler(webapp2.RequestHandler):
+    """Base class for Handlers that render Jinja templates."""
+    def __init__(self, *args, **kwargs):
+        super(RenderingHandler, self).__init__(*args, **kwargs)
+        # The default deadline of 5 seconds is too aggressive of a target for GCS
+        # directory listing operations.
+        urlfetch.set_default_fetch_deadline(60)
+
+    def render(self, template, context):
+        """Render a context dictionary using a given template."""
+        template = JINJA_ENVIRONMENT.get_template(template)
+        self.response.write(template.render(context))
+
+    def check_bucket(self, prefix):
+        if prefix in BUCKET_WHITELIST:
+            return
+        if prefix[:prefix.find('/')] not in BUCKET_WHITELIST:
+            self.abort(404)
